+++ src/main/java/core/gui/model/PlayerColumn2.java
@@ -10,7 +10,7 @@ class PlayerColumn2 extends UserColumn {
 
 	protected PlayerColumn2(int id, String name){
 		super(id,name,name);
-		this.minWidth =20;
+		this.minWidth = 20;
 		preferredWidth = 80;
 	}