+++ src/main/java/module/lineup/LineupTableModel.java
@@ -21,7 +21,7 @@ public class LineupTableModel extends HOTableModel {
 
 	private void initialize() {
 		UserColumn[] basic = UserColumnFactory.createPlayerBasicArray();
-		columns = new UserColumn[49];
+		columns = new UserColumn[50];
 		columns[0] = basic[0];
 		columns[48] = basic[1];
 
@@ -40,7 +40,7 @@ public class LineupTableModel extends HOTableModel {
 		UserColumn[] add = UserColumnFactory.createPlayerAdditionalArray();
 		columns[1] = add[0];
 		columns[2] = add[1];
-		columns[3] = add[11];
+		columns[3] = add[12];
 		columns[4] = add[2];
 		columns[5] = new BooleanColumn(UserColumnFactory.AUTO_LINEUP, " ", "AutoAufstellung", 28);
 		columns[6] = add[4];
@@ -50,6 +50,7 @@ public class LineupTableModel extends HOTableModel {
 		columns[46] = add[7];
 		columns[47] = add[8];
 		columns[22] = add[9];
+		columns[49] = add[10];
 	}
 
 	@Override